backend/requirements.txt
asgiref==3.2.6
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
Django==3.0.4
django-channels==0.7.0
django-rest-framework==0.1.0
django-rest-knox==4.1.0
djangorestframework==3.11.0
idna==2.9
oauthlib==3.1.0
pycparser==2.20
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.8
